+++ bioio_czi/aicspylibczi_reader/reader.py
@@ -26,7 +26,7 @@ from dask import delayed
 from fsspec.implementations.local import LocalFileSystem
 from fsspec.spec import AbstractFileSystem
 
-from .. import metadata_ome as metadata_utils
+from .. import metadata as metadata_utils
 from ..bounding_box import size
 from ..channels import get_channel_names
 from ..pixel_sizes import get_physical_pixel_sizes
