+++ bioio_czi/reader.py
@@ -14,7 +14,7 @@ from ome_types.model.ome import OME
 
 from bioio_czi.aicspylibczi_reader.reader import Reader as AicsPyLibCziReader
 
-from . import metadata_ome
+from . import metadata
 
 
 class Reader(BaseReader):
@@ -218,7 +218,7 @@ class Reader(BaseReader):
             This likely isn't a complete transformation but is guarenteed to
             be a valid transformation.
         """
-        return metadata_ome.transform_metadata_with_xslt(
+        return metadata.transform_metadata_with_xslt(
             self.metadata,
             Path(__file__).parent / "czi-to-ome-xslt/xslt/czi-to-ome.xsl",
         )
